+++ src/fast_diff_py/child_processes.py
@@ -14,7 +14,7 @@ import fast_diff_py.img_processing as imgp
 import fast_diff_py.utils as util
 from fast_diff_py.base_process import GracefulWorker
 from fast_diff_py.cache import BatchCache
-from fast_diff_py.datatransfer_new import PreprocessArg, PreprocessResult, SecondLoopArgs, SecondLoopResults
+from fast_diff_py.datatransfer import PreprocessArg, PreprocessResult, SecondLoopArgs, SecondLoopResults
 
 
 class ChildProcess(GracefulWorker):
